Rebase to the split container images
Because docker hub ran into build timeouts, I split the steps into
separat images. We then noticed, that contrary to the Qt documentation
GCC 4.8.5 is not sufficient for building Qt 5.14. So we switched to
GCC 5.5.0 (the predecessors segfault during Qt compilation) which takes
too long to build on docker hub, too.

So there is not proper reason right now for the splitting, but it had
been done and so just grab the parts from the individual images