Use GCC 4.8.5 and Qt 5.9.9 by default
We also provide GCC 5.5.0 and Qt 5.14.1 that has been build with
GCC 5.5.0.

Qt 5.14.1 needs more work because we need to compile it with
Qt5X11Extras. A requirement is libxkbcommon, which was packaged by Qt
until 5.12.3, then it was removed
https://forum.qt.io/topic/104664/building-qt-5-12-on-centos-6

CentOS6 does not ship libxkbcommon and GCC 4.4.7 fails to compile
libxkbcommon 0.7.1 from CentOS7, so more work is required to be able to
use Qt 5.14 for internal purposes.

Until then make the old build toolchain the default and include a
working but old Qt 5.9.9